MB-29262: Add CompactionConfig struct to compaction_ctx
Hoist the configuration part of the compaction_ctx into its own struct
and pass that struct down the compaction call chain rather than an
entire compaction_ctx.

This is really to simplify the job of adding more to the
compaction_ctx without yet having to ensure all members can be copied
etc... This also defers the creation of the more complex context
structure to the point of invoking compaction.

Change-Id: Ic68ad788e53799e473742fd8181c3ec33c5a214a
Reviewed-on: http://review.couchbase.org/92771
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>