MB-29262: When scanning the seqno index, use the persisted manifest
When a scan is performed we check if items belong to a deleted
collection by calling isLogicallyDeleted, this previously always
used the in-memory manifest to determine the state of the key. This
however is dangerous, particularly when the result of a scan is used
to feed DCP (e.g. backfill_disk).

The in-memory manifest can easily lose updates if a node crashes
allowing a situation when we can backfill a client and drop keys
before a crash, but then after the node restarts a client can stream
and now receive keys which would of been dropped before the crash.

A simple example is a client asking for seqno ranges.

First they ask for all collection keys from seqno 0-10, assume seqno
3 and 4 are for the 'beer' collection which has been deleted from
the memory manifest, but not yet flushed to disk.

* Backfill will send 0 to 9 but drop 3 and 4 (3 is 'create-beer')
* Scan completes and client has received all keys.
* memcached crashes before the beer delete fact has been persisted.
* memcached restarts
* Client reconnects and now asks for 10-19, assume 15 and 16 are
  'beer' mutations
* Backfill will now send every key, none are logically deleted
* Client gets beer keys and has never seen a create-beer or any prior
  'beer' key, an incorrect view of the data.

If the scan drops keys based on the persisted manifest, then this
situation will not occur.

Test-code required some tweaking as we have clash with boost/list.hpp
and ep_test_apis.h both defining check(), this occurred because
callbacks.h now includes vbucket_manifest (and then boost).
callbacks.h has been removed from test_helpers.h

Change-Id: I742e5cccf861baa895f4e72c1ed2b94f2847f4dd
Reviewed-on: http://review.couchbase.org/92774
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>