MB-28615: SCRAM-SHA should not reveal that user doesn't exist
An "attacker" could use SCRAM-SHA and inspect the salt reported
back for a user to determine if a user exists in our system or
not. To get away from that ns_server provides memcached with
a "base salt" to use if the user doesn't exist, then memcached
can generate the actual salt by doing:

    hmac(sha(x), UserName, scramsha_fallback_salt)

This gives a "unique" salt on a per-user base, and a malicious
attacker can't try to connect to memcached to get the salt
and then compare it with the one reported by any other module
in couchbase.

Change-Id: Ie73a76878d9d4cdf93006c8a98583d8fab7cc8e2
Reviewed-on: http://review.couchbase.org/92657
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Jim Walker <jim@couchbase.com>