MB-29227: Add second trigger path for cursor dropping
Add a second trigger condition for dropping checkpoint cursors. This
trigger is if the total memory used for checkpoints is greater than
a % of the bucket quota (initially 50%). If this condition is met,
then we try to free as much memory as we can to take the usage back
down to the lower threshold.

The reason for this change is that we are able to get into a
live-lock like state under the following conditions:

1) All items have been evicted from memory (including meta-data)
2) Checkpoint memory usage is high && overall memory usage >
   high water mark
3) Overall memory usage is < 95%

In these conditions, cursors are not being dropped to free up
memory, but the server is not accepting any more writes to
increase memory usage above the point we start dropping cursors.
The server enters a state where it cannot do anything whilst it
waits for the DCP streams to catch up and free checkpoint memory
through normal operation.

Change-Id: Ia7bc2d9a6a840dfed5972c31308f1942c0e92e30
Reviewed-on: http://review.couchbase.org/92559
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Tim Bradgate <tim.bradgate@couchbase.com>