Add eviction category of stats
To aid with debugging potential performance issues a new category of
eviction stats is created.

In this category we maintain a histogram to record the execution
frequency of items evicted from the hash table (a separate histogram is
used for active/pending vbuckets and replica vbuckets.)

We also maintain a snapshot of the last contents of the frequency
histogram (again, a separate histogram is maintained for active/pending
vbuckets and replica vbuckets.)

The stats can be retrieved using cbstats.  e.g.

./cbstats default:11210 -u xxx -p xxx eviction

The "evicted" histogram stats provide an aggregated view of what the
execution frequencies are for all the items that evicted when running
the hifi_mfu algorithm.

The "snapshot" histogram stats provide a view of what the contents of
the frequency histogram is like during the running of the hifi_mfu
algorithm.

Example output:

 ep_active_or_pending_eviction_values_evicted (102362 total)
    0 - 0   : (  3.78%)  3865 █▊
    1 - 1   : ( 12.99%)  9436 ████▌
    2 - 2   : ( 29.52%) 16913 ████████
    3 - 3   : ( 50.85%) 21832 ██████████▍
    4 - 4   : ( 97.23%) 47477 ██████████████████████▋
    5 - 5   : ( 99.77%)  2602 █▏
    6 - 6   : ( 99.96%)   199
    7 - 7   : (100.00%)    34
    8 - 8   : (100.00%)     3
    10 - 10 : (100.00%)     1
    Avg     : (      3)
 ep_active_or_pending_eviction_values_snapshot (40 total)
    2 - 2     : (  2.50%)  1 █▎
    3 - 3     : ( 22.50%)  8 ██████████
    4 - 4     : ( 47.50%) 10 ████████████▌
    5 - 5     : ( 65.00%)  7 ████████▊
    6 - 6     : ( 72.50%)  3 ███▊
    255 - 255 : (100.00%) 11 █████████████▊
    Avg       : (     73)
 ep_replica_eviction_values_evicted (68829 total)
    0 - 0 : (  1.37%)   945 ▋
    1 - 1 : (  2.01%)   441 ▎
    2 - 2 : (  2.92%)   621 ▍
    3 - 3 : (  6.39%)  2392 █▊
    4 - 4 : (100.00%) 64430 ███████████████████████████████████████████████▋
    Avg   : (      3)
 ep_replica_eviction_values_snapshot (36 total)
    4 - 4     : ( 25.00%)  9 ████████████▌
    255 - 255 : (100.00%) 27 █████████████████████████████████████▌
    Avg       : (    192)

Change-Id: I989efa9624cbbc9aa354c8cefbfef44be4ed258d
Reviewed-on: http://review.couchbase.org/92904
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Tim Bradgate <tim.bradgate@couchbase.com>
Reviewed-by: Jim Walker <jim@couchbase.com>