MB-29274: Add a mutex to ep_test_apis.cc::add_response
A newly enabled test is using many client threads resulting in a race
condition when the shared last_body/last_key/last_ext objects are
written to without a lock.

This patch adds a single lock inside add_response as the tests are not
actually using any of the add_response variables (if they were then
thread local variables will need to be used).


Change-Id: Ife38cced3870f240ec0343e89806ba6797a7a6b4
Reviewed-on: http://review.couchbase.org/92760
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@gmail.com>