MB-29265: Ensure frequencyCounterSaturated callback is initialized
The function invoked when a storedValue's frequency counter becomes
saturated, is held as a callback within the hash table.
It is important that the callback is always intialized to a valid
function otherwise it will throw an exception if accidently invoked.

However this is a "safety net" because when creating vbucket either via
KVBucket::setVBucketState_UNLOCKED or Warmup::createVBuckets we should
set the function to the task responsible for waking the ItemFreqDecayer
task.

The issue was that we did not set the function in the warmup case,
furthermore as we did not initialize at construction it resulted in an
exception being thrown.

This patch addresses the issue by initializing the callback at
construction and also setting it it to the correct value in the warmup
case.

Change-Id: I966ecafa5c58606da16a210a5df8f039832fbd32
Reviewed-on: http://review.couchbase.org/92975
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>