MB-29287: Give each ActiveStream a unique cursor name
To ensure that when the background snapshot processor task runs
and we have closed/created the ActiveStream, if the task gets a handle
on the closed stream we must be sure it does not obtain items destined
for the new stream. Previously with each ActiveStream just using the
name of its producer, the closed stream was able to drain items which
the new stream needed.

Adding a monotonic atomic to the ActiveStream class and appending a
new value to the name we use for the cursor ensures concurrent streams
cannot interfere with each other.

Change-Id: Ie05092490a75c656c344425850eba00043019e96
Reviewed-on: http://review.couchbase.org/93112
Well-Formed: Build Bot <build@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>