MB-29287: Test that takeover finds all items
Using the single threaded test harness we can demonstrate the window
in which a close-stream and create stream will lose items. By
triggering the close/re-create from within the snapshot processor
after it has a handle on the stream which will be closed and before
it access the items, we will transfer items into the dead stream
and prevent the new takeover stream from obtaining them.

Change-Id: I896a7cbd3d32419576294ea88c288b097759a362
Reviewed-on: http://review.couchbase.org/93123
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>