MB-29333: Decay frequency for items considered for eviction
The issue is we are unable to evict when there is only one document in
a vbucket.

The hifi_mfu algorithm will currently only evict the first document it
visits if its execution frequency count is zero.  But documents are
initially added with an initialFreqCount of 4.

The solution is to decrement a document's frequncy count when it visited
but not evicted.  Therefore after sufficient multiple passes of
running the hifi_mfu algorithm the document will be evicted (if it is
eligible).

Change-Id: I95d328238176a97d7ac9e80c268c639bbb647953
Reviewed-on: http://review.couchbase.org/93038
Reviewed-by: Tim Bradgate <tim.bradgate@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>