MB-29381: Allow DCP rollback on vbuckets in pending state
We currently allow rollbacks on only replica vbuckets. This was
added recently (on 4.6.0 with http://review.couchbase.org/#/c/69725).

But during vbucket move, the new master is initially created in
pending state, items streamed over from old master and finally a
takeover stream is created. If there is a rollback during the
takeover, then we should still allow it on the vbucket which in the
pending state.

(Backport of MB-26037 / commit
496d9b96cd046441440022406d561dbc2bd043b5).

Change-Id: Ie57aed02d574be7b8a3852da5a948ef688676770
Reviewed-on: http://review.couchbase.org/93110
Well-Formed: Build Bot <build@couchbase.com>
Reviewed-by: Paolo Cocchi <paolo.cocchi@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@gmail.com>
Tested-by: Build Bot <build@couchbase.com>