MB-29046: Don't crash trying to fetch non-existing response codes
There is no point of transferring the count of the response
codes not used (this just bloat the amout of data being
sent over the wire). If it's not part of the payload, the
number of times its been used is 0.

(This change speeds up the GetSet unit test using Mcbp from
903ms to 690ms, and reduced the StatsCommandContext from
being one of the biggest consumers in my perf output to be
on par with the other CommandContext handlers)

Change-Id: I52c7d2c66133a5aa48dcc4e4116973f7d3c0bf4f
Reviewed-on: http://review.couchbase.org/93159
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Tim Bradgate <tim.bradgate@couchbase.com>