MB-29376: Return Ebusy when audit daemon is backed-up
Currently if the audit daemon event queue gets backed up and reaches
the max_audit_queue size (currently set at 50K) the audit daemon drops
any additional events and reports that the events are dropped in the
memcached log. It then returns AUDIT_FAILED back to memcached.

Memcached then sends an Einternal response back to the client where the
audit event originated from.

go clients interpret the Einternal response as meaning the connection is
unhealthy and therefore drops the connection and hence must reconnect.

The change is for memcached to return Ebusy (which is a more accurate
indication as to the reason why the request was not successful).
go clients can then interpret the Ebusy differently and avoid the
unnecessary disconnect.

Change-Id: I81b2f7f62675029a489e5d31107420a27bed9bab
Reviewed-on: http://review.couchbase.org/93166
Reviewed-by: Tim Bradgate <tim.bradgate@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>