MB-29333: Only decay frequency for items that are eligible for eviction
The original patch did not address the issue correctly.
(see Change-Id: I95d328238176a97d7ac9e80c268c639bbb647953)

It only decrements the frequency count if the item is NOT elgible for
eviction (from the max u8int_t value of 255).  This behaviour is
wrong for two reasons:

1)  We want to decrement the frequency count of items that were eligible
for eviction except that their frequency count is above the threshold.

2)  We don't want to decrement the frequency count that are not eligible
for eviction (regardless of what their frequency count is).

This patch corrects the behaviour to only decrement the fequency count
of items that are eligible for eviction except that their frequency
count is above the threshold.

The original associated test was also found to not exercise the newly
added code.  This has been addressed and an additional test added.

Change-Id: I508964dbf9aefd8055ee7e654caca10d1450aa5b
Reviewed-on: http://review.couchbase.org/93338
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>