MB-29480, MB-29512: Fail backfills that go below purge-seqno
If a backfill is requested and it is not a backfill of everything
the start must be below the purge-seqno, otherwise a DCP client
may miss deletions which have been purged.

This is achieved by loading the purgeSeqno into the ScanContext
and getting backfill create to abort (setting the stream as dead)

The initScanContext method will have opened the data file (and kept
it open) so that the purge-seqno used in the final check won't
change again.

Change-Id: I7505529d46eb6f2b6006695baf7dd9f190237df9
Reviewed-on: http://review.couchbase.org/93690
Reviewed-by: Dave Rigby <daver@couchbase.com>
Reviewed-by: Tim Bradgate <tim.bradgate@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>