MB-29441: DCP Consumer sets correct noop-interval on pre-5.0.0 Producer
Note: the original patch (http://review.couchbase.org/#/c/93911) has been
reverted for MB-29599 (http://review.couchbase.org/#/c/94009). This new
patch contains the fix for MB-29599.

Original commit message:

In MB-19955 we decreased the noop-interval from 180 seconds to 1 second
for DCP Producers. That change is part of versions >=5.0.0.
Note that from MB-19955 a DCP Producer uses the noop-interval only for
sending NOOP messages to the Consumer. That is to implement Fast
Failover.
For detecting dead connections, a post-5.0.0 Producer uses a new
idle-timeout (default value is 360 seconds).

But, on pre-5.0.0 a DCP Producer has a single noop-interval parameter
(default value is 180 seconds), which is used for both sending NOOP
messages to the Consumer and Dead Connection Detection.

When a post-5.0.0 Consumer sets the noop-interval on a pre-5.0.0
Producer (e.g., Online Upgrade with Swap Rebalance, CBSE-5179), it sends
'1 second'. Then the Producer sets 1 second as noop-interval and uses it
for Dead Connection Detection. That makes the Producer to drop all the
connections for which a NOOP response from the Consumer has not arrived
within 1 second.

To fix, we make a Consumer to check the if the Producer is a
pre-5.0.0 and to send the noop-interval accordingly (i.e., 180 seconds
if it is a pre-5.0.0 Producer, 1 second otherwise).
To check the version of the Producer, the Consumer sends a GetErrorMap
request and checks if the command is supported (the GetErrorMap command
is supported from versions >= 5.0.0).

Change-Id: Ie84b69d4943c5c3732509b727ae3b3f0e9893b39
Reviewed-on: http://review.couchbase.org/94055
Reviewed-by: Dave Rigby <daver@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>