getValueAtPercentile to return 0 when histogram is empty
The underlying hdr_histogram cannot store 0 therefore we add 1 before
adding a value to the histogram.  Conversely we minus 1 when we read a
value from the histogram.  Therefore when getValueAtPercentile is called
we minus 1 before returning the value.

However if the histogram is empty the underlying hdr_histogram function
hdr_value_at_percentile returns 0 if the histogram is empty.  Therefore
the getValueAtPercentile returns (0 - 1) which for an unsigned 64-bit
value is the maximum value.

This has been changed so that if the histogram is empty
getValueAtPercentile will return 0, which is the same behaviour as the
underlying hdr_histogram function hdr_value_at_percentile.

Change-Id: Ibc3fedc5ac9127d9fb006ad62b760324ec13051e
Reviewed-on: http://review.couchbase.org/94152
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>