MB-29676: Update memcached bucket flush audit event
ns_server generate an audit event for a bucket flush.  However memcached
generates a similar event for when a memcached binary protocoal command
FLUSH or FLUSHQ is used.

The issue is that for couchbase (and ephemeral) buckets ns_server
performs a bucket flush without using the FLUSH or FLUSHQ commands.
Therefore when the ns_server performs a bucket flush its audit event
appears in the audit log, but the memcached one does not.

On the case where ns_server performs a memcached bucket flush using
the FLUSH command both the ns_server and memcached bucket flush audit
events appear in the log.

To avoid auditing the bucket flush twice, the memcached bucket flush
audit event is only generated if an external user (not ns_server)
issued the flush command.

Change-Id: Id18817ecb6a23cb7b60df35fa89494b324e2c88c
Reviewed-on: http://review.couchbase.org/94209
Reviewed-by: Trond Norbye <trond.norbye@gmail.com>
Tested-by: Build Bot <build@couchbase.com>