MB-29675: Cache is{Snappy,ForceValueCompression}Enabled in ActiveStream
makeResponseFromItem is called for every item to be sent out over DCP;
and it shows up high in 'linux perf' profiles (approx 6% of AuxIO
threads). The bulk of the cost is actually checking if the given
stream supports Snappy, and if all items should be focrcefully
compressed.

This cost comes from both isSnappyEnabled &
isForceValueCompressionEnabled() having to promote the producer
weak_ptr to a shared_ptr to check the relevent producer flag (and then
delete the shared_ptr).

Optimise this by simply caching the value of snappyEnabled /
forceValueCompression in the ActiveStream object at construction time,
and checking the local flag. We don't support changing either of these
flags dynamically for a stream, so this is safe - and avoids all the
shared_ptr manipulation.

With these changes makeResponseFromItem drops to less than 1% of AuxIO
threads.

Tests have been updated as we had one large test which would change
the compression settings of the producer for different scenarios. The
streams compression settings are fixed at the point of streamRequest
so it's easier to split the large test into smaller ones each with
their own producer configured for each scenario

Change-Id: Ice4a559fc7a54bfab4ce9a136d2dc9bdb618e6f4
Reviewed-on: http://review.couchbase.org/94216
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Paolo Cocchi <paolo.cocchi@couchbase.com>