MB-29675: Only wakeup ckpt remover if removable checkpoints
Currently we conditionally wake up the ClosedUnrefCheckpointRemoverTask
in two places, based on if there is more than one checkpoint in
existence:

a) flushVBucket()
b) ActiveStream::getOutstandingItems

However, this is a optimistic check - just because there is more than
one checkpoint; doesn't mean that any checkpoints can actually be
freed. There is typically at least two cursors (persistence +
replication), and potentially many more. Unless the oldest checkpoint is
actually empty (and closed) can we free anything.

As such this will cause us to schedule and run
ClosedUnrefCheckpointRemoverTask unnecessarily.

Change the wakeup criteria to be based on if the oldest checkpoint
(which is the first one which can be considered to be removed) is both
closed, and has zero cursors in it.

Change-Id: I68be804c6d07a991ad53017c9f4fc14ebb9b9a2e
Reviewed-on: http://review.couchbase.org/94296
Reviewed-by: Jim Walker <jim@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Tim Bradgate <tim.bradgate@couchbase.com>