MB-29541: Ensure backfill only takeover will notify the frontend
When a takeover is created and the data is 100% from backfill, we are
missing a notify to the frontend thread as we step from backfill to
take-over send. In that state transition, the code unconditionally
calls nextCheckpointItem which returns false if no data exists in
checkpoints i.e. the stream has no more data. So the frontend thread
can end up sleeping waiting for the stream to notify it of data, but
nothing will come.

Changed the unconditional call to now wake the frontend thread in the
empty case, so that it can step the takeover to its conclusion.

Change-Id: I0e6e3656d1d942b7bf24c86001ec19d5603c173c
Reviewed-on: http://review.couchbase.org/94507
Reviewed-by: Paolo Cocchi <paolo.cocchi@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>