MB-29783: Add tracing extras size to response header correctly
As identified by MB-29558, we do not take into account the size of
the tracing extras when writing to the Pipe if tracing is enabled
for a response header. This leads to us writing to memory not
owned by the buffer of the Pipe object, and hence leads to spurious
crashes of memcahed.

Change-Id: Iedf7b517d8a71f8bff59f2e559f9de7fe72ff6dd
Reviewed-on: http://review.couchbase.org/94599
Reviewed-by: Trond Norbye <trond.norbye@gmail.com>
Tested-by: Build Bot <build@couchbase.com>