MB-29783: Use MCBP_TRACING_RESPONSE_SIZE in mcbp.cc to ensure consistency
As part of the previous patch for this issue, MCBP_TRACING_RESPONSE_SIZE
was introduced as a constant representing the size of the
additional bytes needed fo tracing in the packet header. Use this
constant in mcbp.cc when specifying the tracing extras size to ensure
a consistent view of the world if this value should change in the
future.

Change-Id: Ic8b2d9bacb28185ce326f19608752f908acca11a
Reviewed-on: http://review.couchbase.org/94671
Reviewed-by: Trond Norbye <trond.norbye@gmail.com>
Tested-by: Build Bot <build@couchbase.com>