Add setParam ability for ItemPager configuration parameters
There is now the ability to choose between the original "2-bit_lru" hash
table eviction policy and the new "hifi_mfu" eviction policy.  The
"hifi_mfu" will be the default.

There are a number of configuration paramters associated with the
"hifi_mfu" eviction policy:
- item_eviction_age_percentage
- item_eviction_freq_counter_age_threshold
- item_freq_decayer_chunk_duration
- item_freq_decayer_percent

It is possible that we might need to tune these parameters in the field
and therefore we want to be able to set them at runtime using the
cbepctl tool.

In addition if we find a serious problem with the new "hifi_mfu"
eviction policy, we might want to swap at runtime back to the
"2-bit_lru" policy.

Finally both eviction policies also make use of the following two
configuration parameters:
- pager_active_vb_pcnt
- pager_sleep_time_ms

Therefore it is useful to also ensure that these can be set at runtime.

Change-Id: Ia164ecdfe09e4cef59190b2bf41e1e2bf08aec3b
Reviewed-on: http://review.couchbase.org/94698
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Tim Bradgate <tim.bradgate@couchbase.com>
Reviewed-by: Trond Norbye <trond.norbye@gmail.com>