MB-29915: Provide the ability to switch HT eviction policy at runtime
The hifi_mfu eviction policy is new and has yet to be vigously tested in
"real-life" customer environments.  Therefore we may want to revert to
the original 2-bit_lru policy if a serious issue is found with it.

Given that we currently maintain the counters for the 2-bit_lru policy
and the hfifi_mfu policy, we can switch policies at runtime.  Although
this will not be encouraged (or documented for the end-user) it provides
a fail-safe for Support if the new policy is found to have a severe
adverse effect.

The long term plan is that once we have gained confidence in the new
hifi_mfu eviction policy we will remove the old 2-bit_lru policy and the
ability to switch between the two policies.

Change-Id: Ie6b72ce98ddb05a1aef38468286b7cf9336172d5
Reviewed-on: http://review.couchbase.org/94842
Reviewed-by: Dave Rigby <daver@couchbase.com>
Reviewed-by: Tim Bradgate <tim.bradgate@couchbase.com>
Tested-by: Build Bot <build@couchbase.com>