MB-29675: Add benchmark for StoredValue creation/deletion
To measure the cost of creating/deleting StoredValue objects,
including the memory tracking overhead, create a new microbenchmark
which creates a StoredValue from an Item, then deletes both. Baseline
numbers:

   Run on (8 X 2300 MHz CPU s)
   2018-05-31 12:32:49
   ---------------------------------------------------------------------------------------
   Benchmark                                                Time           CPU Iterations
   ---------------------------------------------------------------------------------------
   VBucketBench/CreateDeleteStoredValue/threads:16        171 ns       1185 ns     685584

Note to ensure thie benchmark is as efficient as possible, the test
fixtures used by these benchmarks have been refactored to only create
a memoryTracker for the specific test cases which require it.

Change-Id: Ie6f99980af84d3cd8eb715b03f8515df297e5570
Reviewed-on: http://review.couchbase.org/94851
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Paolo Cocchi <paolo.cocchi@couchbase.com>