MB-29675: Introduce CoreLocalStats class
To allow additional statistics to be made core-local (in addition to
mem_used), create a new CoreLocalStats class to hold any statistics
which should be tracked on a per-core basis.

This initial patch just moves the existing stat from EPStats to
CoreLocalStats; and then adds an instance of CoreLocalStats to
EPStats.

Change-Id: Iff29f83ba880323f7156e7b26772967a2ee16b27
Reviewed-on: http://review.couchbase.org/94852
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Paolo Cocchi <paolo.cocchi@couchbase.com>
Reviewed-by: Tim Bradgate <tim.bradgate@couchbase.com>