MB-29822: Log disconnections on DCP operations
A disconnection caused by the underlying engine returning
ENGINE_DISCONNECT is currently logged in the proxy function in
engine_erapper.h.
As the DCP executors in memcached do not go through the engine_wrapper.h
functions, we miss logging disconnections caused by ENGINE_DISCONNECT on
DCP operations.
With this patch we redirect all the DCP executors to engine_wrapper.h
and we log the disconnection in there.

Change-Id: I4715dd1a4a33618fa13a770fd2beb4f7efc158a5
Reviewed-on: http://review.couchbase.org/94843
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Tim Bradgate <tim.bradgate@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>