MB-29858: ActiveStream: only request getNumItemsForCursor for existing cursors
ActiveStream::addTakeoverStats() and addStats() both request the
number of items remaining for the Stream's cursor name, even if that
cursor isn't registered (e.g. during backfill). As such, a huge number
of warning log messages are printed in real workloads when streams are
backfilling / have had cursors dropped:

    WARNING (other-2) getNumItemsForCursor_UNLOCKED(): Cursor "<cursor_name>" not found in the checkpoint manager on vb:1013

Remove these messages by only calling getNumItemsForCursor() when a
cursor actually /is/ registered. This is done by introducing a
cursorRegistered flag which is set upon cursor registration and
cleared when cursor removed.

Change-Id: Ib78f073a364f0f7798bfcddfbb0e4eecfcc1790b
Reviewed-on: http://review.couchbase.org/94958
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Tim Bradgate <tim.bradgate@couchbase.com>