Add unit tests for audit generator
This patch adds basic unit tests for the audit generator.
It does not perform negative testing that the parser rejects
incorrect object types in the configuration file, but given
that this tool is used at build time people _SHOULD_ detect
that in the unit test they have for the module using the
new audit descriptors.

Change-Id: Ie1ab77c51056a7cb3c818df4162d68e30f11ab4b
Reviewed-on: http://review.couchbase.org/95023
Tested-by: Build Bot <build@couchbase.com>
Reviewed-by: Dave Rigby <daver@couchbase.com>